+++ crates/fuel-core/src/graphql_api/ports.rs
@@ -1,4 +1,3 @@
-use crate::fuel_core_graphql_api::storage::receipts::Receipts;
 use async_trait::async_trait;
 use fuel_core_services::stream::BoxStream;
 use fuel_core_storage::{
@@ -59,10 +58,22 @@ use fuel_core_types::{
 };
 use std::sync::Arc;
 
-pub trait OffChainDatabase:
-    Send + Sync + StorageInspect<Receipts, Error = StorageError>
-{
+pub trait OffChainDatabase: Send + Sync {
     fn tx_status(&self, tx_id: &TxId) -> StorageResult<TransactionStatus>;
+
+    fn owned_coins_ids(
+        &self,
+        owner: &Address,
+        start_coin: Option<UtxoId>,
+        direction: IterDirection,
+    ) -> BoxedIter<'_, StorageResult<UtxoId>>;
+
+    fn owned_message_ids(
+        &self,
+        owner: &Address,
+        start_message_id: Option<Nonce>,
+        direction: IterDirection,
+    ) -> BoxedIter<'_, StorageResult<Nonce>>;
 
     fn owned_transactions_ids(
         &self,
@@ -73,8 +84,6 @@ pub trait OffChainDatabase:
 }
 
 /// The on chain database port expected by GraphQL API service.
-// TODO: Move `owned_message_ids` and `owned_coins_ids`` to `OffChainDatabase`
-//  https://github.com/FuelLabs/fuel-core/issues/1583
 pub trait OnChainDatabase:
     Send
     + Sync
@@ -86,19 +95,6 @@ pub trait OnChainDatabase:
     + DatabaseChain
     + DatabaseMessageProof
 {
-    fn owned_message_ids(
-        &self,
-        owner: &Address,
-        start_message_id: Option<Nonce>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<Nonce>>;
-
-    fn owned_coins_ids(
-        &self,
-        owner: &Address,
-        start_coin: Option<UtxoId>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<UtxoId>>;
 }
 
 /// Trait that specifies all the getters required for blocks.
@@ -207,7 +203,10 @@ pub mod worker {
             database_description::off_chain::OffChain,
             metadata::MetadataTable,
         },
-        fuel_core_graphql_api::storage::receipts::Receipts,
+        fuel_core_graphql_api::storage::{
+            coins::OwnedCoins,
+            messages::OwnedMessageIds,
+        },
     };
     use fuel_core_services::stream::BoxStream;
     use fuel_core_storage::{
@@ -231,7 +230,8 @@ pub mod worker {
     pub trait OffChainDatabase:
         Send
         + Sync
-        + StorageMutate<Receipts, Error = StorageError>
+        + StorageMutate<OwnedMessageIds, Error = StorageError>
+        + StorageMutate<OwnedCoins, Error = StorageError>
         + StorageMutate<MetadataTable<OffChain>, Error = StorageError>
         + Transactional<Storage = Self>
     {