+++ tests/tests/coin.rs
@@ -2,6 +2,10 @@ use fuel_core::{
     chain_config::{
         CoinConfig,
         StateConfig,
+    },
+    chain_config::{
+        CoinConfig,
+        StateConfig,
         StateReader,
         MAX_GROUP_SIZE,
     },
@@ -33,6 +37,18 @@ use fuel_core_types::{
     fuel_tx::TxId,
 };
 use rstest::rstest;
+
+async fn setup_service(configs: Vec<CoinConfig>) -> FuelService {
+    let mut config = Config::local_node();
+    config.chain_conf.initial_state = Some(StateConfig {
+        coins: Some(configs),
+        ..Default::default()
+    });
+
+    FuelService::from_database(Database::default(), config)
+        .await
+        .unwrap()
+}
 
 #[tokio::test]
 async fn coin() {
@@ -75,27 +91,15 @@ async fn first_5_coins(
 
     // setup test data in the node
     let coins: Vec<_> = (1..10usize)
-        .map(|i| Coin {
-            utxo_id: UtxoId::new([i as u8; 32].into(), 0),
+        .map(|i| CoinConfig {
             owner,
             amount: i as Word,
-            asset_id: Default::default(),
-            maturity: Default::default(),
-            tx_pointer: Default::default(),
+            ..Default::default()
         })
         .collect();
 
-    let mut db = Database::default();
-    for coin in coins {
-        db.storage::<Coins>()
-            .insert(&coin.utxo_id.clone(), &coin.compress())
-            .unwrap();
-    }
-
     // setup server & client
-    let srv = FuelService::from_database(db, Config::local_node())
-        .await
-        .unwrap();
+    let srv = setup_service(coins).await;
     let client = FuelClient::from(srv.bound_address);
 
     // run test
@@ -122,27 +126,16 @@ async fn only_asset_id_filtered_coins() {
 
     // setup test data in the node
     let coins: Vec<_> = (1..10usize)
-        .map(|i| Coin {
-            utxo_id: UtxoId::new([i as u8; 32].into(), 0),
+        .map(|i| CoinConfig {
             owner,
             amount: i as Word,
             asset_id: if i <= 5 { asset_id } else { Default::default() },
-            maturity: Default::default(),
-            tx_pointer: Default::default(),
+            ..Default::default()
         })
         .collect();
 
-    let mut db = Database::default();
-    for coin in coins {
-        db.storage::<Coins>()
-            .insert(&coin.utxo_id.clone(), &coin.compress())
-            .unwrap();
-    }
-
     // setup server & client
-    let srv = FuelService::from_database(db, Config::local_node())
-        .await
-        .unwrap();
+    let srv = setup_service(coins).await;
     let client = FuelClient::from(srv.bound_address);
 
     // run test
@@ -170,28 +163,17 @@ async fn get_coins_forwards_backwards(
     #[values(AssetId::from([1u8; 32]), AssetId::from([32u8; 32]))] asset_id: AssetId,
 ) {
     // setup test data in the node
-    let coins: Vec<Coin> = (1..11usize)
-        .map(|i| Coin {
-            utxo_id: UtxoId::new([i as u8; 32].into(), 0),
+    let coins: Vec<_> = (1..11usize)
+        .map(|i| CoinConfig {
             owner,
             amount: i as Word,
             asset_id,
-            maturity: Default::default(),
-            tx_pointer: Default::default(),
+            ..Default::default()
         })
         .collect();
 
-    let mut db = Database::default();
-    for coin in coins {
-        db.storage::<Coins>()
-            .insert(&coin.utxo_id.clone(), &coin.compress())
-            .unwrap();
-    }
-
     // setup server & client
-    let srv = FuelService::from_database(db, Config::local_node())
-        .await
-        .unwrap();
+    let srv = setup_service(coins).await;
     let client = FuelClient::from(srv.bound_address);
 
     // run test